Moving Error to package com.cdancy.bitbucket.rest.domain.common
